Implement workaround for turbo + scroll smooth behavior
If user has scrolled down the page, then navigates to a new page, Turbo
will scroll back to the top of the page. We use this workaround to
disable scroll-behavior: smooth on the <html> element before the visit,
then re-enable it after the visit so that the smooth scroll won't run on
Turbo visit. This won't affect users who haven't scrolled or who
otherwise do not have scroll-behavior set.

Credit to @randoum for issue comment:
https://github.com/turbolinks/turbolinks/issues/181#issuecomment-727549337