Mount liteboard app on routes
Unfortunately, the addition of litestack brings in dependencies on
hanami-router which is locked to rack 2.x. This means downgrading from
rack 3.x to rack 2.x—or otherwise fore-going the use of the liteboard
web interface.

For now, I've downgraded rack by making the dependency on 2.x explicit
in the Gemfile. I've also opened a new issue on hanami-router to track
progress toward Rack 3 support: https://github.com/hanami/router/issues/268