Add Pwa::InstallationInstructions model
The purpose of this model is to convert a user agent string into a
partial name for OS + Browser combinations for which we can provide
PWA installation info.

A list of user agent strings is provided by https://www.useragents.me/

We use the DeviceDetector gem for user agent string parsing.