Build custom litestream.yml on boot for litestream-ruby
litestream-ruby assumes a litestream.yml in config/ but it is not
flexible enough to support multiple configured dbs for replication such
as the data, cache, metrics, and queue databases used for litestack. We
take an approach here where Rails will generate the environment-specific
litestream.yml using Rails credentials on boot. This will work for
production well. It has the unfortunate behavior currently though that
running development and test on the same application will overwrite the
config file.