Disable Flipper preload
Flipper inserts middleware that preloads all Flipper flags into memory.
By default this is run on every request. Since we're using flags
sparingly at the moment, we temporarily disable this behavior
completely. The middleware supports options to control what requests the
flags are preloaded on—this may be helpful should we choose to revisit
at a later date.