Migrate from jest to vitest
Vitest integrates more closely with Vite and provides some powerful
features beyond Jest. It's mostly Jest-compatible, which made this an
easy switch.

Of note:

- Vite and Vitest have started issuing warnings that the CJS build of
  Vite is deprecated. To address these warnings, we added "type":
"module" to pacakge.json and update the postcss and tailwindcss config
files to ESM exports from CommonJS module.exports.
- Globals are disabled by default in Vitest. We added imports for
  expect/describe/it etc. to the test utils; however, we rely on
@testing-library/jest-dom for matcher extensions and this library
appears to expect "expect" to be globally defined. To address, we've
added globals: true to the vitest config anyways.