Remove litestack
Having replaced litedb with sqlite adapter, litecache with solid cache,
and litequeue with solid queue, it's increasingly hard to justify
keeping litestack. I've found it difficult to configure and not as
robust as I would like. This changeset removes all litestack
dependencies and references.

A key missing piece is a Sqlite-backed Action Cable adapter. There's an
open issue in Rails that suggests this may be a part of Rails 8 in the
future: https://github.com/rails/rails/issues/50480. For now, we fall
back to the default `async` adapter for action cable which will have to
do for the time being.