Add admin index page
When signing in as an admin, we'll know redirect to the admin index
page, which contains links to the admin engines like jobs and flipper
flags. While itself useful, this also is added to side-step the fact
that Site pages will now have HTTP caching, so flash messages will not
work out-of-the-box on these pages. We could potentially do something
with turbo + flash, but punt on that for now.