Disable rack mini profiler when caching enabled
HTTP caching wasn't working in development and it was due to Rack Mini
Profiler.

From the Rack mini profiler README:

> To fix some nasty bugs with rack-mini-profiler showing the wrong data,
the middleware will remove headers relating to caching (Date & Etag on
responses, If-Modified-Since & If-None-Match on requests). This probably
won't ever break your application, but it can cause some unexpected
behavior. For example, in a Rails app, calls to stale? will always
return true.

https://github.com/MiniProfiler/rack-mini-profiler?tab=readme-ov-file#caching-behavior