Use hard-coded uuid for atom entry id
Post records may be associated with Sitepress articles in the future.
For now, the hard-coded uuid will act as temporary solution for
providing a unique identifier for atom feed entry content. Any
modification to the url could make the content appear to be new for feed
readers so that may be something to account for in future article
changes.