Add User model
We prevent empty values from being saved into the email column through a
null: false constraint in addition to the presence validation.

We enforce unique email addresses at the database level through
add_index :users, :email, unique: true in addition to a uniqueness
validation.

We ensure all emails are valid through a format validation.

We save all emails to the database in a downcase format via a
before_save callback such that the values are saved in a consistent
format.

We use URI::MailTo::EMAIL_REGEXP that comes with Ruby to validate that
the email address is properly formatted.

https://stevepolito.design/blog/rails-authentication-from-scratch#step-1-build-user-model