Configure user to generate confirmation token
The has_secure_password method is added to give us an API to work with
the password_digest column.

The confirm! method will be called when a user confirms their email
address. We still need to build this feature.

The confirmed? and unconfirmed? methods allow us to tell whether a user
has confirmed their email address or not.

The generate_confirmation_token method creates a signed_id that will be
used to securely identify the user. For added security, we ensure that
this ID will expire. This will be useful when we build the confirmation
mailer.

https://stevepolito.design/blog/rails-authentication-from-scratch#step-2-add-confirmation-and-password-columns-to-users-table

Fix for add password to user