Add email confirmation
The create action will be used to resend confirmation instructions to an
unconfirmed user. We still need to build this mailer, and we still need
to send this mailer when a user initially signs up. This action will be
requested via the form on app/views/confirmations/new.html.erb. Note
that we call downcase on the email to account for case sensitivity when
searching.

The edit action is used to confirm a user’s email. This will be the page
that a user lands on when they click the confirmation link in their
email. We still need to build this. Note that we’re looking up a user
through the find_signed method and not their email or ID. This is
because The confirmation_token is randomly generated and can’t be
guessed or tampered with unlike an email or numeric ID. This is also why
we added param: :confirmation_token as a named route parameter.

https://stevepolito.design/blog/rails-authentication-from-scratch#step-3-create-sign-up-pages