Add confirmation mailer
The send_confirmation_email! method will create a new
confirmation_token. This is to ensure confirmation links expire and
cannot be reused. It will also send the confirmation email to the user.

We call update_columns so that the updated_at/updated_on columns are not
updated. This is personal preference, but those columns should typically
only be updated when the user updates their email or password.

The links in the mailer will take the user to
ConfirmationsController#edit at which point they’ll be confirmed.

https://stevepolito.design/blog/rails-authentication-from-scratch#step-5-create-confirmation-mailer