Add current user methods
The Current class inherits from ActiveSupport::CurrentAttributes which
allows us to keep all per-request attributes easily available to the
whole system. In essence, this will allow us to set a current user and
have access to that user during each request to the server.

The Authentication Concern provides an interface for logging the user in
and out. We load it into the ApplicationController so that it will be
used across the whole application.

We use warden to serialize the user’s ID in the session so that we can
have access to the user across requests. The user’s ID won’t be stored
in plain text. The cookie data is cryptographically signed to make it
tamper-proof. And it is also encrypted so anyone with access to it can’t
read its contents.

The redirect_if_authenticated method checks to see if the user is logged
in. If they are, they’ll be redirected to the root_path. This will be
useful on pages an authenticated user should not be able to access, such
as the login page.

The current_user method returns a User and sets it as the user on the
Current class we created. We use memoization to avoid fetching the User
each time we call the method. We call the before_action filter so that
we have access to the current user before each request. We also add this
as a helper_method so that we have access to current_user in the views.

The user_signed_in? method simply returns true or false depending on
whether the user is signed in or not. This is helpful for conditionally
rendering items in views.

https://stevepolito.design/blog/rails-authentication-from-scratch#step-6-create-current-model-and-authentication-concern