Add ability to log in as user
Adds a Sessions controller and views for new sessions

The create method simply checks if the user exists and is confirmed. If
they are, then we check their password. If the password is correct, we
log them in via the login method we created in the Authentication
Concern. Otherwise, we render an alert.

We’re able to call user.authenticate because of has_secure_password

Note that we call downcase on the email to account for case sensitivity
when searching.

Note that we set the flash to “Incorrect email or password.” if the user
is unconfirmed. This prevents leaking email addresses.

The destroy method simply calls the logout method we created in the
Authentication Concern.

The login form is passed a scope: :user option so that the params are
namespaced as params[:user][:some_value]. This is not required, but it
helps keep things organized.

https://stevepolito.design/blog/rails-authentication-from-scratch#step-7-create-login-page