Add password reset email
Creates a signed token that will be used to securely identify the user.
For added security, we ensure that this ID will expire in 10 minutes
(this can be controlled with the PASSWORD_RESET_TOKEN_EXPIRATION
constant) and give it an explicit purpose of :reset_password.

Create a new password_reset_token. This is to ensure password reset links expire and cannot be reused. It will also send the password reset email to the user.

https://stevepolito.design/blog/rails-authentication-from-scratch#step-9-add-password-reset-functionality