Add model for changing email
We add a unconfirmed_email table so that we have a place to store the
email a user is trying to use after their account has been confirmed
with their original email.

We ensure to format the unconfirmed_email before saving it to the
database. This ensures all data is saved consistently.

We add validations to the unconfirmed_email table ensuring it’s a valid
email address.

We update the confirm! method to set the email column to the value of
the unconfirmed_email table, and then archive all pending unconfirmed
emails. This will only happen if a user is trying to confirm a new email
address. Note that we return false if updating the email address fails.
This could happen if a user tries to confirm an email address that has
already been confirmed.

We add the confirmable_email method so that we can call the correct
email in the updated UserMailer.

We add reconfirming? and unconfirmed_or_reconfirming? to help us
determine what state a user is in. This will come in handy later in our
controllers.