Add ability to update email or password
The authenticate_user! method can be called to ensure an anonymous user
cannot access a page that requires a user to be logged in. We’ll need
this when we build the page allowing a user to edit or delete their
profile.

We call authenticate_user! before editing, destroying, or updating a
user since only an authenticated user should be able to do this.

We update the create method to accept create_user_params (formerly
user_params). This is because we’re going to require different
parameters for creating an account vs. editing an account.

The destroy action simply deletes the user and logs them out. Note that
we’re calling current_user, so this action can only be scoped to the
user who is logged in.

The edit action simply assigns @user to the current_user so that we have
access to the user in the edit form.

The update action first checks if their password is correct. Note that
we’re passing this in as current_password and not password. This is
because we still want a user to be able to change their password and
therefore we need another parameter to store this value. This is also
why we have a private update_user_params method.

If the user is updating their email address (via unconfirmed_email) we
send a confirmation email to that new email address before setting it as
the email value.

We force a user to always put in their current_password as an extra
security measure in case someone leaves their browser open on a public
computer.

We disable the email field to ensure we’re not passing that value back
to the controller. This is just so the user can see what their current
email is.

We require the current_password field since we’ll always want a user to
confirm their password before making changes.

The password and password_confirmation fields are there if a user wants
to update their current password.

Adapted from
https://stevepolito.design/blog/rails-authentication-from-scratch#step-12-update-users-controller