Add friendly login redirects
The store_location method stores the request.original_url in the session
so it can be retrieved later. We only do this if the request made was a
get request. We also call request.local? to ensure it was a local
request. This prevents redirecting to an external application.

We call store_location in the authenticate_user! method so that we can
save the path to the page the user was trying to visit before they were
redirected to the login page. We need to do this before visiting the
login page otherwise the call to request.original_url will always return
the url to the login page.

The after_login_path variable it set to be whatever is in the
session[:user_return_to]. If there’s nothing in session[:user_return_to]
then it defaults to the root_path.

Note that we call this method before calling login. This is because
login calls reset_session which would deleted the
session[:user_return_to].

Adapted from
https://stevepolito.design/blog/rails-authentication-from-scratch#step-17-add-friendly-redirects