Use ULID for user primary key
BREAKING CHANGE: This change breaks all sorts of rules by editing
migrations. This would not be normally recommended but we’re allowing an
exception for this new feature which is still behind a feature flag.

Adds the sqlite-ulid gem
Updates database.yml to enable support for the new extension
Modifies existing migration to make user primary key ULIDs and updates
the type for user foreign keys

Credit for implementation to
https://fractaledmind.github.io/2023/09/22/enhancing-rails-sqlite-ulid-primary-keys/