Add invisible captcha for registration and newsletter forms
Extracts a module for using the invisible captcha view helper with
Phlex. Since the helper makes uses of "concat" and "session" I also had
to use the Phlex Rails macro convention for registering access to those
helper methods.

We also need to deal with invisible captcha for specs; currently
disabling timing and spinner checks.

We also need to deal with the gem for wasm because we reference it in an
initializer.