Move example post to new action, handle redirect from frame
Now UX is to click a "New post" button that will render the post form
dynamically in a target turbo frame. The Stimulus refresh controller is
renamed as it now needs to handle a special redirect use case; on
successful submission of the POST from the "examples_post_form" we want
to render an updated list in the "examples_posts" turbo frame. The
framework does not yet provide an intuitive mechanism for breaking out
of a frame on redirect, so we use a Stimulus Turbo.visit to the
"examples_post" form here. This does result in a "double GET" request,
the first results in re-rendering the requesting "examples_post_form",
but the second is needed to re-render the "examples_posts" frame.

There are other approaches, described in the resources below, which also
lay out the problem in more detail:

- https://www.ducktypelabs.com/turbo-break-out-and-redirect/
- https://github.com/hotwired/turbo-rails/pull/367#issuecomment-1601733561
- https://discuss.hotwired.dev/t/break-out-of-a-frame-during-form-redirect/1562/26