Support upload of base64 image data to snippet with active storage
Use Stimulus to capture 'turbo:before-fetch-request' event to modify the
request body. This event allows for pausing and resuming the form submit
event on the Snippet form asynchronously. This enables us to use the
html-to-image library to render the Snippet html as a base64 data uri
image and append to the form submit request body. On the backend, we
enable the attachment by converting the base64 data to image data and
attaching to the Snippet.