Upgrade solid cache to address NoMethodError
SolidCache is raising errors in production following Rails 7.2 upgrade.

NoMethodError - undefined method `model' for class SolidCache::Entry

Same as reported here:
https://github.com/rails/solid_cache/issues/189

The error was not captured in specs because we have been using the
NullStore for caching in tests.

This error can be reproduced by setting SolidCacheStore in tests. I
don’t want to enable this as the backend right now but may reconsider if
we continue to see issues across upgrades.