Convert combobox ux to a11y-improved turbo stream
Improves a11y by implementing aria-expanded and aria-activedescendant in
the SearchCombobox Stimulus controller.

Converts the Turbo Frame interaction to a Turbo Stream for more
flexibility. This is similar to how Hotwire Combobox renders async
options.

Identifies Listbox options by id so we can set the aria-activedescendant
attribute when the active selection changes.

Adds a SearchListbox Stimulus controller to communicate that the Turbo
Stream has rendered so the Combobox can update the aria-expanded
attribute.