Enhance the search combobox open/close behavior
This changeset improves the interactions for openingn and closing the
search combobox within the dialog.

- Adds open/close methods to change a new stimulus value "expandedValue"
- On change of "expandedValue", the aria-expanded value is changed and
  one of expand/collapse is called
- expand/collapse displays or hides the listbox, respectively
- adds cross-controller communication so when the dialog is opened or
  closed, the combobox open or close methods are called.

More:

Some browsers clear the focused search element when the ESC key is
pressed. now the search element won’t be cleared by the dialog will
close.

Now, when the dialog is closed, the current query is retained, so when
the dialog opens again, the search query is still present. I kind of
like this behavior, but it may not be the behavior some users expect.
Let’s see, maybe no one cares enough.