Refactor Vector serialization to work better with AR
sqlite-vec will convert a JSON.encoded array for us on insert. We want
to avoid binary-unpack on deserializing this instance because the
Ruby-representation will still be the JSON encoded array. After the
record is loaded from the database however, the "before type cast" value
will be the binary representation, at which point, we can use
String#unpack. Accounting for this allows us to replace the
"custom_create" method.