Implement Page#related_pages as association
After a bunch of experimentation here’s the best approach I’ve come up
with so far to be able to treat the concept of "similar_pages" as an
association on Page.

We’re using a subquery to find the embeddings that match the embedding
of the given page. We then join the pages table to the subquery to get
the related pages. We order by the distance between the embeddings and
exclude the page itself.

We unscope(:where) because we want to disable the default association
behavior of scoping where(id: page.id) for querying related pages not
including the page itself.

We use the associated scope to ensure that the page_embedding
association exists when querying the related pages, otherwise, the MATCH
embedding subquery will raise an ActiveRecord::StatementInvalid error;
there needs to be an embedding to match against!