Forward ruby enumeration demo params with dark mode
A bit convoluted here, but continuing down this path of keeping the demo
JS in a separate repo: we want to request an iframe URL with query
params with color scheme choices. Dark mode is not kept on the server at
the moment, so I use a Stimulus controller to set the final turbo frame
url on the frontend which is later converted to the appropriate iframe
url on the backend.

Send query params from Ruby to JS to Ruby to JS!!!